REFACT: Module structure
- dedicated module for specifications; and
- split large tests into smaller files.
