+++ ip_validation/specifications/struct_reqs.py
